hotel/requirements.txt
streamlit
opencv-python
python-dotenv
requests